Feedback implementations

+++ app/models/Roles.js
@@ -1,7 +1,11 @@
-import Validator from './../helpers/validate';
+import Validator from '../helpers/Validate';
 
-module.exports = (sequelize, DataTypes) => {
-  const roles = sequelize.define('roles', {
+/**
+ * Roles models
+ * @class Roles
+ */
+const RoleModel = (sequelize, DataTypes) => {
+  const Roles = sequelize.define('Roles', {
     title: DataTypes.STRING,
   }, {
     classMethods: {
@@ -9,28 +13,35 @@ module.exports = (sequelize, DataTypes) => {
         // associations can be defined here
       },
 
+      /**
+       * Creates a new role entry
+       * @method createRole
+       * @param {String} title
+       * @return Promise
+       */
       createRole: (title) => {
         const validate = new Validator();
 
         return new Promise((fulfill, fail) => {
           if (!validate.isEmpty(title)) {
-            roles.roleExists(title).then(() => {
-              fail({
-                statusCode: 403,
-                message: 'Role already exists',
-              });
-            })
-              .catch(() => {
-                roles.create({
-                  title,
-                }).then((role) => {
-                  fulfill(role);
-                }).catch((error) => {
-                  fail({
-                    statusCode: 500,
-                    message: error,
-                  });
+            Roles.roleExists(title)
+              .then(() => {
+                fail({
+                  statusCode: 403,
+                  message: 'Role already exists',
                 });
+              })
+              .catch(() => {
+                Roles.create({ title })
+                  .then((role) => {
+                    fulfill(role);
+                  })
+                  .catch((error) => {
+                    fail({
+                      statusCode: 500,
+                      message: error,
+                    });
+                  });
               });
           } else {
             fail({
@@ -41,13 +52,15 @@ module.exports = (sequelize, DataTypes) => {
         });
       },
 
+      /**
+       * Retrieves a specific role
+       * @method getRole
+       * @param {String, int} id
+       * @return Promise
+      */
       getRole: (id) => {
         return new Promise((fulfill, fail) => {
-          roles.find({
-            where: {
-              id,
-            },
-          })
+          Roles.find({ where: { id } })
             .then((role) => {
               if (role) {
                 fulfill(role.dataValues);
@@ -61,31 +74,40 @@ module.exports = (sequelize, DataTypes) => {
         });
       },
 
+      /**
+       * Asserts if a role exists
+       * @method roleExists
+       * @param {String} title
+       * @return Promise
+       */
       roleExists: (title) => {
         return new Promise((fulfill, fail) => {
-          roles.find({
-            where: {
-              title,
-            },
-          })
+          Roles.find({ where: { title } })
             .then((role) => {
               if (role) {
                 fulfill(role.dataValues);
               }
 
               fail('Role does not exists');
-            }).catch((error) => {
+            })
+            .catch((error) => {
               fail(error);
             });
         });
       },
 
+      /**
+       * Retrieves all roles in the system
+       * @method all
+       * @return Promise
+       */
       all: () => {
         return new Promise((fulfill, fail) => {
-          roles.findAll()
+          Roles.findAll()
             .then((allRoles) => {
               fulfill(allRoles);
-            }).catch((error) => {
+            })
+            .catch((error) => {
               fail({
                 statusCode: 500,
                 message: error,
@@ -94,23 +116,23 @@ module.exports = (sequelize, DataTypes) => {
         });
       },
 
+      /**
+       * Updates the a specific role with new details
+       * @method updateRole
+       * @param {String, int} roleId
+       * @param {String} title
+      */
       updateRole: (roleId, title) => {
         return new Promise((fulfill, fail) => {
-          roles.find({
-            where: {
-              id: roleId,
-            },
-          })
+          Roles.find({ where: { id: roleId } })
             .then((role) => {
-              role.updateAttributes({
-                title,
-              })
+              role.updateAttributes({ title })
                 .then((newRole) => {
                   fulfill(newRole);
                 })
-              .catch((error) => {
-                fail(error);
-              });
+                .catch((error) => {
+                  fail(error);
+                });
             })
             .catch((error) => {
               fail(error);
@@ -118,23 +140,21 @@ module.exports = (sequelize, DataTypes) => {
         });
       },
 
+      /**
+       * Delete a role from the database
+       * @method deleteRole
+       * @param {String, int} roleId
+      */
       deleteRole: (roleId) => {
         return new Promise((fulfill, fail) => {
-          roles.find({
-            where: {
-              id: roleId,
-            },
-          })
+          Roles.find({ where: { id: roleId } })
             .then((role) => {
               if (role) {
-                role.destroy({
-                  where: {
-                    id: roleId,
-                  },
-                })
+                role.destroy({ where: { id: roleId } })
                   .then(() => {
                     fulfill('Role deleted');
-                  }).catch((error) => {
+                  })
+                  .catch((error) => {
                     fail({
                       statusCode: 500,
                       message: error,
@@ -157,5 +177,7 @@ module.exports = (sequelize, DataTypes) => {
       },
     },
   });
-  return roles;
+  return Roles;
 };
+
+export default RoleModel;
